Add a symmetric set of tables to represent files in the control silo. This basically duplicates the File model graph by abstracting the base functionality into django abstract models and realizing concrete models for both silos. This approach is jankier than I would like it to be, so I'm definitely open to suggestion on the approach.
