+++ migrations_lockfile.txt
@@ -6,5 +6,5 @@ To resolve this, rebase against latest master and regenerate your migration. Thi
 will then be regenerated, and you should be able to merge without conflicts.
 
 nodestore: 0002_nodestore_no_dictfield
-sentry: 0457_sentry_monitorcheckin_date_added_index
+sentry: 0458_control_files
 social_auth: 0001_initial
