+++ src/sentry/models/files/abstractfileblobindex.py
@@ -1,0 +1,10 @@
+from sentry.db.models import BoundedPositiveIntegerField, Model
+
+
+class AbstractFileBlobIndex(Model):
+    __include_in_export__ = False
+
+    offset = BoundedPositiveIntegerField()
+
+    class Meta:
+        abstract = True
