+++ src/sentry/models/files/abstractfileblobowner.py
@@ -1,0 +1,10 @@
+from sentry.db.models import BoundedBigIntegerField, Model
+
+
+class AbstractFileBlobOwner(Model):
+    __include_in_export__ = False
+
+    organization_id = BoundedBigIntegerField(db_index=True)
+
+    class Meta:
+        abstract = True
